divide unit test with module
