+++ include/aquarius/elastic/archive.hpp
@@ -32,9 +32,9 @@ namespace aquarius
 			{
 				this->roll_back();
 
-				//_Ty error{};
+				_Ty error{};
 
-				//std::swap(t, error);
+				std::swap(t, error);
 
 				result = false;
 			}
