+++ test/Test.cpp
@@ -1,11 +1,5 @@
 #define BOOST_TEST_MODULE UnitTest
 #include <boost/test/included/unit_test.hpp>
-#include "io.h"
-#include "message_process.h"
-#include "attr.h"
-#include "generate.h"
-#include "inherit.h"
-#include "type.h"
-#include "tuple_size.h"
-#include "els.h"
-#include "io_service_pool_test.h"
+#include "server_test.h"
+#include "elastic_test.h"
+#include "message_test.h"
