+++ test/message_test.h
@@ -1,0 +1,1 @@
+#pragma once
